requirements.txt
discord.py
python-dotenv
sqlitedict
boto3
